:recycle: AddReadyにlockかける, ファイル分割
