+++ usecases/service/ws/client.go
@@ -130,73 +130,3 @@ func (c *Client) readPump() {
 		}
 	}
 }
-
-func (c *Client) bloadcast(next func(c *Client)) {
-	for _, m := range c.room.Members {
-		cc, ok := c.hub.userIdToClient[m.Id]
-		if !ok {
-			continue
-		}
-
-		go next(cc)
-	}
-}
-
-func (c *Client) sendMsg(msg []byte) {
-	// TODO: unregisterできるようにする
-	if c.send == nil {
-		if c.userId == c.room.HostId {
-			c.sendChangeHostEvent()
-		}
-		c.hub.unregister(c)
-
-		return
-	}
-
-	c.send <- msg
-}
-
-func (c *Client) sendMsgToEachClientInRoom(msg []byte) {
-	c.bloadcast(func(cc *Client) {
-		cc.sendMsg(msg)
-	})
-}
-
-func (c *Client) allMembersAreReady() bool {
-	c.hub.mux.Lock()
-	defer c.hub.mux.Unlock()
-
-	r := c.room
-	for _, m := range r.Members {
-		if _, ok := c.hub.userIdToClient[m.Id]; !ok {
-			continue
-		}
-
-		if _, ok := r.Game.Ready[m.Id]; !ok {
-			return false
-		}
-	}
-
-	return true
-}
-
-func (c *Client) resetTimer() {
-	// タイマーをリセットする
-	// 15(分)後に次のゲームが始まらなければルームを削除する
-	game := c.room.Game
-	bt := game.BreakTimer
-
-	if stopped := bt.Stop(); !stopped {
-		go c.waitAndBreakRoom()
-	}
-
-	bt.Reset(time.Minute * 15)
-	go c.waitAndBreakRoom()
-}
-
-func (c *Client) waitAndBreakRoom() {
-	<-c.room.Game.BreakTimer.C
-	if err := c.sendBreakRoomEvent(); err != nil {
-		log.Println("failed to send BREAK_ROOM event:", err.Error())
-	}
-}
